Moved existing code for OSVVM projects to pyEDAA.OSVVM.Project.
