+++ pyEDAA/OSVVM/__init__.py
@@ -33,7 +33,7 @@ __author__ =    "Patrick Lehmann"
 __email__ =     "Paebbels@gmail.com"
 __copyright__ = "2025-2025, Patrick Lehmann"
 __license__ =   "Apache License, Version 2.0"
-__version__ =   "0.4.0"
+__version__ =   "0.5.0"
 __keywords__ =  ["eda project", "model", "abstract", "osvvm", "reports", "data format", "unit test", "test suite", "test bench", "test harness"]
 
 from pyTooling.Decorators import export
